use 1H structure as soft model feature; remove 1H gate from structure labels
- features.py: expose htf_1h_structure as continuous float32 model input (67th
  feature, Group 9); backbone can learn when 1H alignment matters without hard
  filtering the label set
- labels.py: generate_structure_labels now uses forward asymmetry only (>1.5
  bullish, <0.67 bearish); no longer gated by _1h_structure — expected to raise
  confirmed label density from ~9% to ~30-40%
- config.py: num_features 66 → 67
- model.py: logits.float() before loss computation to prevent AMP float16
  overflow → NaN → KeyError: 'loss'
- colabs/ffm_pretrain_5min.py: bfloat16 auto-detection (A100 skips GradScaler),
  torch.amp.GradScaler new API, clarify val split message (50% sequences is
  expected with stride=4 train; 20% bars is what matters)
- tests: 175 pass — update feature count 66→67, structure label tests no longer
  require struct_1h fixture, add htf_1h_structure column test

+++ futures_foundation/model.py
@@ -401,7 +401,7 @@ class FFMForPretraining(PreTrainedModel):
 
         for task_name, labels, logits, loss_fn, weight in labels_and_logits:
             if labels is not None:
-                task_loss = loss_fn(logits, labels)
+                task_loss = loss_fn(logits.float(), labels)  # fp32 for stable loss under AMP
                 # CrossEntropyLoss returns nan when every sample in the batch is the
                 # ignore_index (all-sentinel batch for a masked head). Skip that head
                 # so the combined loss stays finite — semantically identical to not
